refactor(core): move lockfile and tool DTOs into cuenv-manifest (2d)
- Platform/Os/Arch, ToolSource, ToolExtract, and the tool activation
  step DTOs move from core::tools into cuenv_manifest::tools; core
  re-exports them and keeps the provider trait, registry, and
  activation resolution logic.
- The full lockfile schema (Lockfile, LockedTool, LockedToolPlatform,
  LockedVcsDependency, artifacts, runtimes) moves to
  cuenv_manifest::lockfile with its own LockfileError; load/save keep
  their behavior and core::Error gains a From<LockfileError> impl so
  existing `?` call sites compile unchanged. The canonical
  to_tool_source conversion from Phase 1.3 travels with the types.

Also disables incremental compilation artifacts in this environment
(the 20G incremental cache filled the disk); no repo-visible change.

Validation: workspace check --all-targets clean; cuenv-manifest (77
tests incl. moved lockfile suite), cuenv-ci (334), cuenv (433) green,
core lib green minus the known sandbox-only process-tree kill test;
clippy clean on touched crates under CI-pinned lints; cargo fmt; git
diff --check.
