+++ crates/manifest/Cargo.toml
@@ -13,6 +13,8 @@ categories.workspace = true
 [dependencies]
 serde = { workspace = true }
 serde_json = { workspace = true }
+thiserror = { workspace = true }
+toml = { workspace = true }
 cuenv-hooks = { workspace = true }
 cuenv-task-graph = { workspace = true }
